+++ demos/js-webshim/minified/extras/modernizr-custom.js
@@ -1,5 +1,0 @@
-if(window.console && console.error){
-	console.error('"js-webshim" folder was moved from http://afarkas.github.io/webshim/demos/js-webshim/ to http://afarkas.github.io/webshim/js-webshim/. Please update your path');
-} else {
-	alert('webshim has moved please update your path');
-}