+++ demos/js-webshim/minified/polyfiller.js
@@ -1,1 +1,0 @@
-alert('"js-webshim" folder was moved from http://afarkas.github.io/webshim/demos/js-webshim/ to http://afarkas.github.io/webshim/js-webshim/. Please update your path');