+++ changelog.md
@@ -1,4 +1,4 @@
-upcomming Release 1.13.2
+Release 1.14.0
 ----------
 - implement ``inputmode="numeric"`` for Smartphones ([see demo](http://fiddle.jshell.net/trixta/7NEBb/show/))
 - re-implemented FileReader